+++ lib/cgpt/lib/foundation/et.h
@@ -41,7 +41,8 @@ void cgpt_set_to_number(Lattice<T> & l, ComplexD val) {
   autoView(l_v, l, AcceleratorWriteDiscard);
   auto p_v = &l_v[0];
   constexpr int n_elements = GridTypeMapper<T>::count;
-  accelerator_for(ss, grid->oSites() * n_elements, grid->Nsimd(), {
+  
+  accelerator_for(ss, grid->oSites() * n_elements, (size_t)grid->Nsimd(), {
       auto osite = ss / n_elements;
       auto j = ss - osite * n_elements;
       coalescedWriteElement(p_v[osite], v, j);
